Moving whitespace tests into Tests.cs.
