+++ SgmlTests/Tests.cs
@@ -17,6 +17,8 @@ using NUnit.Framework;
 using Sgml;
 using System.Xml.Linq;
 using System.Xml.XPath;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace SGMLTests {
 
@@ -442,6 +444,67 @@ namespace SGMLTests {
 
         }
 
-        
+        /// <summary>This is a unit-test for the <see cref="SgmlReader.TextWhitespace"/> property's logic.</summary>
+        [Test]
+        public void Invalid_TextWhitespace_enum_values_should_be_dropped()
+        {
+            SgmlReader sgmlReader = new SgmlReader()
+            {
+                TextWhitespace = (TextWhitespaceHandling)0xFF
+            };
+
+            Assert.AreEqual(TextWhitespaceHandling.TrimBoth | TextWhitespaceHandling.OnlyLineBreaks, sgmlReader.TextWhitespace, "The " + nameof(sgmlReader.TextWhitespace) + " property should only respect defined flags bits.");
+        }
+
+        /// <summary>https://github.com/lovettchris/SgmlReader/issues/15</summary>
+        [Test]
+        public void DoBradescoTrailingWhitespaceTest()
+        {
+            // 63.test is originally "bradesco.ofx" from github.com/kevencarneiro/OFXSharp, used under the MIT license.
+            string bradescoOfxText = Tests.ReadTestResource(name: "63.test");
+            bradescoOfxText = bradescoOfxText.Substring(startIndex: 134); // skip first 134 chars of OFX header.
+
+            SgmlDtd ofx160Dtd = Tests.LoadDtd(docType: "OFX", name: "ofx160.dtd");
+
+            XmlDocument xmlDocument;
+            using (StringReader stringReader = new StringReader(bradescoOfxText))
+            {
+                SgmlReader sgmlReader = new SgmlReader()
+                {
+                    InputStream        = stringReader,
+                    WhitespaceHandling = WhitespaceHandling.None,
+                    TextWhitespace     = TextWhitespaceHandling.OnlyTrailingLineBreaks,
+                    DocType            = ofx160Dtd.Name,
+                    Dtd                = ofx160Dtd
+                };
+
+                Assert.AreEqual(TextWhitespaceHandling.OnlyTrailingLineBreaks, sgmlReader.TextWhitespace, "The " + nameof(sgmlReader.TextWhitespace) + " property should persist this valid value.");
+
+                xmlDocument = new XmlDocument();
+                using (XmlWriter xmlWriter = xmlDocument.CreateNavigator().AppendChild())
+                {
+                    while (!sgmlReader.EOF)
+                    {
+                        xmlWriter.WriteNode(sgmlReader, defattr: true);
+                    }
+                }
+            }
+
+            // Assert:
+            XmlNode codeNode = xmlDocument.SelectSingleNode("//*[local-name()='CODE']");
+            XmlNodeList memoNodes = xmlDocument.SelectNodes("//*[local-name()='MEMO']");
+
+            List<XmlNode> selectedNodes = new List<XmlNode>();
+            selectedNodes.Add(codeNode);
+            selectedNodes.AddRange(memoNodes.Cast<XmlNode>());
+
+            foreach (XmlNode node in selectedNodes)
+            {
+                string innerText = codeNode.InnerText;
+                bool hasTrailingLineBreak = innerText.EndsWith("\n");
+
+                Assert.IsFalse(hasTrailingLineBreak, message: "There should be no trailing line-breaks in <" + node.Name + "> elements' innerText.");
+            }
+        }
     }
 }
