+++ SgmlTestsCore/SgmlTestsCore.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <Import Project="../Common/version.props" />
 
   <PropertyGroup>
@@ -92,7 +92,8 @@
     <EmbeddedResource Include="..\SgmlTests\Resources\59.test" Link="Resources\59.test" />
     <EmbeddedResource Include="..\SgmlTests\Resources\60.test" Link="Resources\60.test" />
     <EmbeddedResource Include="..\SgmlTests\Resources\61.test" Link="Resources\61.test" />
-	<EmbeddedResource Include="..\SgmlTests\Resources\62.test" Link="Resources\62.test" />
+    <EmbeddedResource Include="..\SgmlTests\Resources\62.test" Link="Resources\62.test" />
+    <EmbeddedResource Include="..\SgmlTests\Resources\63.test" Link="Resources\63.test" />
   </ItemGroup>
 
 </Project>
